ControlsHelper split up
