+++ MahApps.Metro/Styles/Clean/CleanGroupBox.xaml
@@ -5,7 +5,7 @@
     <Style TargetType="GroupBox" 
            x:Key="CleanGroupBoxStyleKey">
         <Setter Property="BorderThickness" Value="0.3" />
-        <Setter Property="Controls:ControlsHelper.HeaderFontSize"
+        <Setter Property="Controls:TabControlHelper.HeaderFontSize"
                 Value="16" />
         <Setter Property="Template">
             <Setter.Value>
@@ -17,9 +17,9 @@
                         </Grid.RowDefinitions>
 
                         <ContentPresenter Margin="{TemplateBinding Padding}"
-                                          TextElement.FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
-                                          TextElement.FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
-                                          TextElement.FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
+                                          TextElement.FontSize="{TemplateBinding Controls:TabControlHelper.HeaderFontSize}"
+                                          TextElement.FontStretch="{TemplateBinding Controls:TabControlHelper.HeaderFontStretch}"
+                                          TextElement.FontWeight="{TemplateBinding Controls:TabControlHelper.HeaderFontWeight}"
                                           ContentSource="Header"
                                           ContentTemplate="{TemplateBinding HeaderTemplate}"
                                           RecognizesAccessKey="True"
