+++ MahApps.Metro/Styles/Controls.Scrollbars.xaml
@@ -572,7 +572,7 @@
 
     <Style x:Key="MetroScrollViewer"
            TargetType="{x:Type ScrollViewer}">
-        <Setter Property="Controls:ControlsHelper.VerticalScrollBarOnLeftSide"
+        <Setter Property="Controls:TabControlHelper.VerticalScrollBarOnLeftSide"
                 Value="False" />
         <Setter Property="Template">
             <Setter.Value>
@@ -621,7 +621,7 @@
                                    ViewportSize="{TemplateBinding ViewportWidth}" />
                     </Grid>
                     <ControlTemplate.Triggers>
-                        <Trigger Property="Controls:ControlsHelper.VerticalScrollBarOnLeftSide"
+                        <Trigger Property="Controls:TabControlHelper.VerticalScrollBarOnLeftSide"
                                  Value="True">
                             <Setter TargetName="leftColumn"
                                     Property="Width"
