+++ MahApps.Metro/Themes/MetroAnimatedSingleRowTabControl.xaml
@@ -9,7 +9,7 @@
     </ResourceDictionary.MergedDictionaries>
 
     <Style TargetType="{x:Type Controls:MetroAnimatedSingleRowTabControl}">
-        <Setter Property="Controls:ControlsHelper.Transition"
+        <Setter Property="Controls:TabControlHelper.Transition"
                 Value="Left" />
         <Setter Property="Template">
             <Setter.Value>
@@ -39,7 +39,7 @@
                                       KeyboardNavigation.TabIndex="1" />
                         </ScrollViewer>
                         <Controls:TransitioningContentControl x:Name="ContentPanel"
-                                                              Transition="{TemplateBinding Controls:ControlsHelper.Transition}"
+                                                              Transition="{TemplateBinding Controls:TabControlHelper.Transition}"
                                                               Behaviours:ReloadBehavior.OnSelectedTabChanged="True"
                                                               RestartTransitionOnContentChange="True"
                                                               Grid.Column="0"
@@ -71,7 +71,7 @@
                         </Trigger>
                         <Trigger Property="TabStripPlacement"
                                  Value="Left">
-                            <Setter Property="Controls:ControlsHelper.Transition"
+                            <Setter Property="Controls:TabControlHelper.Transition"
                                     Value="Right" />
                             <Setter Property="Grid.Row"
                                     TargetName="HeaderPanelScroll"
@@ -136,7 +136,7 @@
                             <Setter Property="HorizontalScrollBarVisibility"
                                     TargetName="HeaderPanelScroll"
                                     Value="Disabled" />
-                            <Setter Property="Controls:ControlsHelper.VerticalScrollBarOnLeftSide"
+                            <Setter Property="Controls:TabControlHelper.VerticalScrollBarOnLeftSide"
                                     TargetName="HeaderPanelScroll"
                                     Value="True" />
                         </Trigger>
