+++ MahApps.Metro/Themes/MetroTabItem.xaml
@@ -32,7 +32,7 @@
         <Setter Property="SnapsToDevicePixels"
                 Value="True" />
         <!-- special property for header font size -->
-        <Setter Property="Controls:ControlsHelper.HeaderFontSize"
+        <Setter Property="Controls:TabControlHelper.HeaderFontSize"
                 Value="{DynamicResource TabItemFontSize}" />
         <Setter Property="Template">
             <Setter.Value>
@@ -48,16 +48,16 @@
                                               Margin="2,1,2,1"
                                               HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
                                               VerticalAlignment="{TemplateBinding VerticalContentAlignment}"
-                                              TextElement.FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
-                                              TextElement.FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
-                                              TextElement.FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
+                                              TextElement.FontSize="{TemplateBinding Controls:TabControlHelper.HeaderFontSize}"
+                                              TextElement.FontStretch="{TemplateBinding Controls:TabControlHelper.HeaderFontStretch}"
+                                              TextElement.FontWeight="{TemplateBinding Controls:TabControlHelper.HeaderFontWeight}"
                                               TextElement.Foreground="{TemplateBinding Foreground}"
                                               ContentSource="Header"
                                               RecognizesAccessKey="True" />
                             <Button x:Name="PART_CloseButton"
                                     Style="{DynamicResource ChromelessButtonStyle}"
                                     Background="{DynamicResource GrayNormalBrush}"
-                                    Width="{TemplateBinding Controls:ControlsHelper.HeaderFontSize, Converter={StaticResource MetroTabItemCloseButtonWidthConverter}}"
+                                    Width="{TemplateBinding Controls:TabControlHelper.HeaderFontSize, Converter={StaticResource MetroTabItemCloseButtonWidthConverter}}"
                                     Height="{Binding RelativeSource={RelativeSource Self}, Path=Width, Mode=OneWay}"
                                     HorizontalAlignment="Right"
                                     VerticalAlignment="Top"
